requirements.txt
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
Django==2.0
django-cors-headers==2.4.0
django-pagination==1.0.7
django-phonenumber-field==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.0
passlib==1.7.1
pexpect==4.6.0
phonenumberslite==8.9.9
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
